+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add rectangles to Annotation, add position to referee, add actions to JobDescriptionSearchConfig
 
 ## [2.6.0] - 2023-01-09
 ### Fixed